[QFEED-124] Rename: Answer API 계층 DTO 디렉토리 추가
