+++ module-api/src/main/java/com/wsws/moduleapi/feed/dto/answer/get/AnswerListGetApiResponse.java
@@ -1,6 +1,5 @@
-package com.wsws.moduleapi.feed.dto.answer;
+package com.wsws.moduleapi.feed.dto.answer.get;
 
-import com.wsws.moduleapplication.feed.dto.answer.AnswerFindServiceResponse;
 import com.wsws.moduleapplication.feed.dto.answer.AnswerListFindServiceResponse;
 
 import java.util.List;
