+++ module-api/src/main/java/com/wsws/moduleapi/feed/dto/answer/post/AnswerPostApiRequest.java
@@ -1,4 +1,4 @@
-package com.wsws.moduleapi.feed.dto.answer;
+package com.wsws.moduleapi.feed.dto.answer.post;
 
 import com.wsws.moduleapplication.feed.dto.answer.AnswerCreateServiceRequest;
 import jakarta.validation.constraints.NotNull;
