+++ module-api/src/main/java/com/wsws/moduleapi/feed/dto/answer/post/AnswerPostApiResponse.java
@@ -1,4 +1,4 @@
-package com.wsws.moduleapi.feed.dto.answer;
+package com.wsws.moduleapi.feed.dto.answer.post;
 
 
 import com.wsws.moduleapplication.feed.dto.answer.AnswerCreateServiceResponse;
